+++ grunt/config/less.js
@@ -15,8 +15,7 @@ module.exports = {
 			sourceMap: true
 		},
 		files  : {
-			'<%= paths.app %>/styles/css/hobbynote-test-app.css': '<%= paths.app %>/styles/less/hobbynote-test-app.less',
-			'<%= paths.app %>/styles/css/init-loader.css'       : '<%= paths.app %>/styles/less/hobbynote-test-app.loader.less'
+			'<%= paths.app %>/styles/css/hobbynote-test-app.css': '<%= paths.app %>/styles/less/hobbynote-test-app.less'
 		}
 	},
 	tmp : {
@@ -25,8 +24,7 @@ module.exports = {
 			sourceMap: true
 		},
 		files  : {
-			'.tmp/release/styles/hobbynote-test-app.css': '<%= paths.app %>/styles/less/hobbynote-test-app.less',
-			'.tmp/release/styles/init-loader.css'       : '<%= paths.app %>/styles/less/hobbynote-test-app.loader.less'
+			'.tmp/release/styles/hobbynote-test-app.css': '<%= paths.app %>/styles/less/hobbynote-test-app.less'
 		}
 	}
 };