+++ CHANGELOG.md
@@ -10,9 +10,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Copy fonts grunt task for release build
 - New view `main`
 - New provider `urlRouterProvider` to configure the default route
+- Webicon bower dependency
 
 ### Changed
 - Refactored ES6 stuff to avoid issues
 
 ### Removed
 - Useless stuff
+- Init loader stuff